ui/GitHub-Summarizer-QuestionGenerator/requirements.txt
# REQUIRES PYTHON 3.11+
# This project uses openai-agents which requires Python 3.11 or higher

# Core API dependencies
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pydantic>=1.8.0
python-dotenv>=0.19.0

# OpenAI agents framework - REQUIRES PYTHON 3.11+
openai-agents>=0.1.0

# OpenAI API
openai>=1.0.0

# Google Gemini API
google-generativeai>=0.3.0

# HTTP client
requests>=2.28.0
httpx>=0.24.0

# Additional utilities
python-multipart>=0.0.5

# Note: If you have Python 3.9, you'll need to upgrade to Python 3.11+
# You can check your Python version with: python --version
